fix undefined user state and rename action types
