+++ src/components/Login.js
@@ -4,7 +4,7 @@ import {
   Container, Row, Col, Form, Button,
 } from 'react-bootstrap';
 import { Link, useNavigate } from 'react-router-dom';
-import { login, setErrors } from '../redux/user/userSessions';
+import { login, setErrors } from '../redux/userSessions/userSessions';
 
 function Login() {
   const navigate = useNavigate();
