+++ src/components/Logout.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { logout } from '../redux/user/userSessions';
+import { logout } from '../redux/userSessions/userSessions';
 
 function Login() {
   const navigate = useNavigate();
