+++ src/redux/configureStore.js
@@ -1,11 +1,11 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import user from './user/userSessions';
+import userSessions from './userSessions/userSessions';
 import reservations from './reservations/reservations';
 import vehicles from './vehicles/vehicles';
 
-const reducer = combineReducers({ user, vehicles, reservations });
+const reducer = combineReducers({ userSessions, vehicles, reservations });
 
 const store = createStore(
   reducer,
